backend/requirements.txt
livekit-agents==0.8.0
livekit-api==0.5.0
livekit==0.11.0
google-genai==0.3.0
mem0ai==0.1.5
python-dotenv==1.0.0
asyncio
uvloop